requirements.txt
numpy==1.22.2
matplotlib==3.5.1
streamlit==1.12.0
requests==2.26.0
PyMuPDF==1.19.6
Pillow==9.2.0
plotly